data-science-assistant-Editeds/requirements.txt
streamlit
pandas
matplotlib
seaborn
scikit-learn
sqlalchemy
python-dotenv
qwen-agent